Poprawiona grafika (zrobiona postac playera). Na poczatku aplikacji dodany loadbar (nic nie robiacy ;) ). Menu obsluguje tylko wlaczenie gry i wyjscie.

+++ include/SpriteConfig.hpp
@@ -29,7 +29,7 @@ private:
 		} 
       
       void Insert(const string& name, ushort tex_x, ushort tex_y, short pos_x,
-                  short pos_y, ushort width,ushort height, uint frame_count,
+                  short pos_y, ushort width,ushort height, ushort frame_count,
                   float frame_time ,bool loop ){  				  	  
 		 SpriteData tmp={tex_x,tex_y,pos_x,pos_y,width,height,frame_count,frame_time,loop};
 		_data.insert(std::make_pair(name, tmp));
